requirements.txt
openai>=1.14.0
aiohttp>=3.9.3
pydantic>=2.6.4
pydantic-settings>=2.2.1
chromadb>=0.4.24
aiosqlite>=0.20.0
python-telegram-bot>=21.0.1
discord.py>=2.3.2
slack-bolt>=1.18.1
rich>=13.7.1
click>=8.1.7
apscheduler>=3.10.4
ollama>=0.1.7
PyYAML>=6.0.1
textual>=0.52.1
SQLAlchemy>=2.0.28
